Move testing utilities to separate class.
