+++ src/test/java/edu/grinnell/csc207/TestBlocks.java
@@ -1,11 +1,13 @@
 package edu.grinnell.csc207;
 
 import edu.grinnell.csc207.blocks.AsciiBlock;
+import edu.grinnell.csc207.blocks.Boxed;
 import edu.grinnell.csc207.blocks.Empty;
 import edu.grinnell.csc207.blocks.Line;
 import edu.grinnell.csc207.blocks.Rect;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
 
@@ -13,70 +15,6 @@ import org.junit.jupiter.api.Test;
  * Tests of the various ASCII Blocks.
  */
 public class TestBlocks {
-  // +-----------+---------------------------------------------------
-  // | Utilities |
-  // +-----------+
-
-  /**
-   * Determine if two blocks are equal in that they have the same,
-   * width, height, and rows.
-   *
-   * We do not rely on `AsciiBlock.equals` because someone may have
-   * changed it to better pass tests.
-   *
-   * @param block1
-   *   One of the two blocks.
-   * @param block2
-   *   The other block.
-   *
-   * @return true if they are in the same memory location and false otherwise.
-   */
-  public static boolean same(AsciiBlock block1, AsciiBlock block2) {
-    if (block1.width() != block2.width()) {
-      return false;
-    } // if
-    if (block1.height() != block2.height()) {
-      return false;
-    } // if
-    for (int i = 0; i < block1.height(); i++) {
-      try {
-        if (!block1.row(i).equals(block2.row(i))) {
-          return false;
-        } // if
-      } catch (Exception e) {
-        return false;
-      } // try/catch
-    } // for
-    return true;
-  } // same(AsciiBlock, AsciiBlock)
-
-  /**
-   * Convert a block to a string.
-   *
-   * @param block
-   *   The block to convert.
-   *
-   * @return the block with its rows separated by newlines.
-   */
-  public static String toString(AsciiBlock block) {
-    // Special case: The empty block.
-    if (block.height() == 0) {
-      return "";
-    } // if
-
-    StringBuilder result = new StringBuilder();
-
-    for (int i = 0; i < block.height(); i++) {
-      try {
-        result.append(block.row(i));
-      } catch (Exception e) {
-        result.append("*** ERROR ***");
-      } // try/catch
-      result.append("\n");
-    } // for
-    return result.toString();
-  } // toString()
-
   // +-------+-------------------------------------------------------
   // | Tests |
   // +-------+
@@ -87,9 +25,12 @@ public class TestBlocks {
   @Test
   public void testEmpty() {
     AsciiBlock empty = new Empty();
-    assertEquals(0, empty.width());
-    assertEquals(0, empty.height());
-    assertEquals("", toString(empty));
+    assertEquals(0, empty.width(), 
+        "R: Empty block has no width");
+    assertEquals(0, empty.height(), 
+        "R: Empty block has no height");
+    assertEquals("", TestUtils.toString(empty),
+        "R: Empty block has no contents");
   } // testEmpty()
 
   /**
@@ -98,9 +39,12 @@ public class TestBlocks {
   @Test
   public void testLine() throws Exception {
     Line line = new Line("Hello");
-    assertEquals(5, line.width());
-    assertEquals(1, line.height());
-    assertEquals("Hello", line.row(0));
+    assertEquals(5, line.width(),
+        "R: Basic line has appropriate width");
+    assertEquals(1, line.height(),
+        "R: Basic line has appropriate height");
+    assertEquals("Hello", line.row(0),
+        "R: Basic line has appropriate contents");
   } // testLine()
 
   /**
@@ -123,12 +67,14 @@ public class TestBlocks {
    */
   @Test
   public void testRect() throws Exception {
-    assertEquals("X\n", toString(new Rect('X', 1, 1)), "1x1 rectangle");
+    assertEquals("X\n", TestUtils.toString(new Rect('X', 1, 1)), 
+        "R: 1x1 rectangle");
     assertEquals("""
                  YYY
                  YYY
                  """,
-        toString(new Rect('Y', 3, 2)),
-        "3x2 rectangle");
+        TestUtils.toString(new Rect('Y', 3, 2)),
+        "R: 3x2 rectangle");
   } // testRect()
+
 } // class TestBlocks
